requirements.txt

aiogram
aiohttp
fastapi
uvicorn
jinja2
python-dotenv
certifi
